Update SklearnWorker.py
When self.fraction=1, which means to use the entire train_idx. But if sampling randomly, the order of samples will be changed, which means their corresponding samples are the same, but the order is different. This will affect model like random forest, which is sensitive to the sample order. So, I suggest that when self.fraction=1, train_idx should be used directly instead of randomly sampling all the train_idx. This is the source code problem of sklearn,  and I also submitted suggestions to them.